requirements.txt
tensorflow==2.9.*
keras==2.9.*
numpy==1.22.*
pandas==1.4.*
pillow==9.1.*
